backend/pyproject.toml
[tool.poetry]
name = "content-hub"
version = "0.1.0"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"

fastapi = "^0.115.10"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
pydantic = {extras = ["email"], version = "^2.10.6"}
pydantic-settings = "^2.8.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.38"}
asyncpg = "^0.30.0"
alembic = "^1.15.1"
pillow = "^11.2.1"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

